% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExtra-package.R
\docType{package}
\name{MultiAssayExtra-package}
\alias{MultiAssayExtra}
\alias{MultiAssayExtra-package}
\title{MultiAssayExtra: Testing package for additional features}
\description{
Test feature requests.
}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

}
\keyword{internal}
